+++ examples/common/logs_foo_library/foo_library.cc
@@ -32,7 +32,7 @@ void foo_library()
   auto scoped_span = trace::Scope(get_tracer()->StartSpan("foo_library"));
   auto ctx         = span->GetContext();
   auto logger      = get_logger();
-  logger->Log(opentelemetry::logs::Severity::kDebug, "name", "body", {}, {}, ctx.trace_id(),
+  logger->Log(opentelemetry::logs::Severity::kDebug, "name", "body", {}, ctx.trace_id(),
               ctx.span_id(), ctx.trace_flags(), opentelemetry::common::SystemTimestamp());
 }
 #endif