+++ exporters/otlp/include/opentelemetry/exporters/otlp/otlp_log_recordable.h
@@ -40,63 +40,61 @@ public:
    * Set the timestamp for this log.
    * @param timestamp the timestamp to set
    */
-  virtual void SetTimestamp(opentelemetry::common::SystemTimestamp timestamp) noexcept override;
+  void SetTimestamp(opentelemetry::common::SystemTimestamp timestamp) noexcept override;
 
   /**
    * Set the severity for this log.
    * @param severity the severity of the event
    */
-  virtual void SetSeverity(opentelemetry::logs::Severity severity) noexcept override;
+  void SetSeverity(opentelemetry::logs::Severity severity) noexcept override;
 
   /**
    * Set name for this log
    * @param name the name to set
    */
-  virtual void SetName(nostd::string_view name) noexcept override;
+  void SetName(nostd::string_view name) noexcept override;
 
   /**
    * Set body field for this log.
    * @param message the body to set
    */
-  virtual void SetBody(nostd::string_view message) noexcept override;
+  void SetBody(nostd::string_view message) noexcept override;
 
   /**
-   * Set a single resource of a log record.
-   * @param key the name of the resource to set
-   * @param value the resource value to set
+   * Set Resource of this log
+   * @param Resource the resource to set
    */
-  virtual void SetResource(nostd::string_view key,
-                           const opentelemetry::common::AttributeValue &value) noexcept override;
+  void SetResource(const opentelemetry::sdk::resource::Resource &resource) noexcept override;
 
   /**
    * Set an attribute of a log.
    * @param key the name of the attribute
    * @param value the attribute value
    */
-  virtual void SetAttribute(nostd::string_view key,
-                            const opentelemetry::common::AttributeValue &value) noexcept override;
+  void SetAttribute(nostd::string_view key,
+                    const opentelemetry::common::AttributeValue &value) noexcept override;
 
   /**
    * Set the trace id for this log.
    * @param trace_id the trace id to set
    */
-  virtual void SetTraceId(opentelemetry::trace::TraceId trace_id) noexcept override;
+  void SetTraceId(opentelemetry::trace::TraceId trace_id) noexcept override;
 
   /**
    * Set the span id for this log.
    * @param span_id the span id to set
    */
-  virtual void SetSpanId(opentelemetry::trace::SpanId span_id) noexcept override;
+  void SetSpanId(opentelemetry::trace::SpanId span_id) noexcept override;
 
   /**
    * Inject trace_flags for this log.
    * @param trace_flags the trace flags to set
    */
-  virtual void SetTraceFlags(opentelemetry::trace::TraceFlags trace_flags) noexcept override;
+  void SetTraceFlags(opentelemetry::trace::TraceFlags trace_flags) noexcept override;
 
 private:
   proto::logs::v1::LogRecord log_record_;
-  opentelemetry::sdk::common::AttributeMap resource_attributes_;
+  const opentelemetry::sdk::resource::Resource *resource_ = nullptr;
   // TODO shared resource
   // const opentelemetry::sdk::resource::Resource *resource_ = nullptr;
   // TODO InstrumentationLibrary