+++ sdk/test/metrics/meter_provider_sdk_test.cc
@@ -5,6 +5,8 @@
 #  include <gtest/gtest.h>
 #  include "opentelemetry/sdk/metrics/meter.h"
 #  include "opentelemetry/sdk/metrics/meter_provider.h"
+#  include "opentelemetry/sdk/metrics/view/instrument_selector.h"
+#  include "opentelemetry/sdk/metrics/view/meter_selector.h"
 
 using namespace opentelemetry::sdk::metrics;
 
@@ -36,15 +38,14 @@ public:
   bool Shutdown() noexcept override { return true; }
 };
 
-class MockView : public View
-{};
-
 TEST(MeterProvider, GetMeter)
 {
   std::vector<std::unique_ptr<MetricExporter>> exporters;
   std::vector<std::unique_ptr<MetricReader>> readers;
-  std::vector<std::unique_ptr<View>> views;
-  MeterProvider mp1(std::move(exporters), std::move(readers), std::move(views));
+
+  MeterProvider mp1(std::move(exporters), std::move(readers));
+  //   std::unique_ptr<View> view{std::unique_ptr<View>()};
+  // MeterProvider mp1(std::move(exporters), std::move(readers), std::move(views);
   auto m1 = mp1.GetMeter("test");
   auto m2 = mp1.GetMeter("test");
   auto m3 = mp1.GetMeter("different", "1.0.0");
@@ -76,7 +77,11 @@ TEST(MeterProvider, GetMeter)
   std::unique_ptr<MetricReader> reader{new MockMetricReader()};
   ASSERT_NO_THROW(mp1.AddMetricReader(std::move(reader)));
 
-  std::unique_ptr<View> view{new MockView()};
-  ASSERT_NO_THROW(mp1.AddView(std::move(view)));
+  std::unique_ptr<View> view{std::unique_ptr<View>()};
+  std::unique_ptr<InstrumentSelector> instrument_selector{
+      new InstrumentSelector(InstrumentType::kCounter, "instru1")};
+  std::unique_ptr<MeterSelector> meter_selector{new MeterSelector("name1", "version1", "schema1")};
+  ASSERT_NO_THROW(
+      mp1.AddView(std::move(instrument_selector), std::move(meter_selector), std::move(view)));
 }
 #endif